src/modalchess/utils/config.py
"""YAML 설정 로딩 유틸리티."""

from __future__ import annotations

from pathlib import Path
from typing import Any

import yaml


def load_yaml_config(path: str | Path) -> dict[str, Any]:
    """YAML 설정 파일을 딕셔너리로 읽는다."""
    with Path(path).open("r", encoding="utf-8") as handle:
        data = yaml.safe_load(handle)
    return data or {}
