pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "modalchess"
version = "0.1.0"
description = "Spatial chess research core for ModalChess."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "python-chess>=1.999",
  "torch>=2.2",
  "transformers>=4.40",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
modalchess-train = "modalchess.train.train_spatial_baseline:main"
modalchess-eval = "modalchess.eval.eval_baseline:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
